requirements.txt
aiogram==3.22.0
apscheduler==3.9.1
reportlab==3.6.12
matplotlib==3.7.1
python-binance==1.0.16
pandas>=2.3.2
uvicorn==0.23.2
fastapi==0.103.2
python-dotenv==1.0.1
anyio<4.0.0,>=3.7.1
appnope==0.1.4 ; sys_platform == 'darwin'
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0 ; python_full_version >= '3.14'
argon2-cffi-bindings==25.1.0 ; python_full_version < '3.14'
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel>=2.9.1,<2.10.0
beautifulsoup4==4.13.5
bleach==6.2.0
bokeh==3.7.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6 ; sys_platform == 'win32'
comm==0.2.3
contourpy==1.3.3
cryptography==45.0.6
curl-cffi==0.13.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.0
fastjsonschema==2.21.2
fonttools==4.59.1
fqdn==1.5.1
frozendict==2.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.30.1
ipython==9.4.0
ipython-pygments-lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lark==1.2.2
llvmlite==0.44.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.3
multitasking==0.0.12
narwhals==2.2.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandocfilters==1.5.1
parso==0.8.5
peewee==3.18.2
pexpect==4.9.0 ; sys_platform != 'emscripten' and sys_platform != 'win32'
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
prometheus-client==0.22.1
prompt-toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
ptyprocess==0.7.0 ; os_name != 'nt' or (sys_platform != 'emscripten' and sys_platform != 'win32')
pure-eval==0.2.3
pycparser==2.22
pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytz==2025.2
pywin32==311 ; platform_python_implementation != 'PyPy' and sys_platform == 'win32'
pywinpty==3.0.0 ; os_name == 'nt'
pyyaml==6.0.2
pyzmq==27.0.2
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.0
send2trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250822
typing-extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xyzservices==2025.4.0
yfinance==0.2.65
pandas-ta==0.4.71b0
aiohttp>=3.9.0,<3.13